+++ FileReportDlg.cpp
@@ -214,7 +214,7 @@ void CFileReportDlg::OnBrowse()
 void CFileReportDlg::OnPeriodChange()
 {
 	// Get current selection.
-	int i = m_cbPeriod.CurSel();
+	size_t i = m_cbPeriod.CurSel();
 	ASSERT(i != CB_ERR);
 
 	// Convert selelction to date range.
