+++ PrintReportDlg.cpp
@@ -211,7 +211,7 @@ void CPrintReportDlg::OnSelect()
 void CPrintReportDlg::OnPeriodChange()
 {
 	// Get current selection.
-	int i = m_cbPeriod.CurSel();
+	size_t i = m_cbPeriod.CurSel();
 	ASSERT(i != CB_ERR);
 
 	// Convert selelction to date range.
